added inital support for scene updates
scene updates have been missing from the plugin, most likely for a good
reason: It's not entirely clear what the status of a scene should be
(and the ISY REST service has no state for it either). The practical
approach used here is that if any device in a scene has a value other
than DOF/0, the the scene is considered ON. If all devices are at level
DOF/0 then the scene is OFF as well.